requirements.txt
openai==1.13.3
tiktoken==0.6.0
cohere==4.53
anthropic==0.18.1
replicate==0.22.0
octoai-sdk==0.7.4
monsterapi==1.0.5.post1
together==0.2.9
fireworks-ai==0.9.0
mistralai==0.0.8
python-dotenv==1.0.0
pre-commit==3.6.0
lamini==2.0.4
groq==0.4.1
# Future Providers (maybe)
# boto3==1.33.10
